ruff: enforce unnecessary-construct + cleanliness tranche (bd fafs)
Graduate RUF005/007/010/022/046, C408/C420, SIM105/108/300, PERF401, PLW0108,
E731 to enforced. Real fixes (13 autofixed + the rest by hand):
- SIM105: tracking.py telemetry try/except/pass -> contextlib.suppress (x4)
- E731: query.py/synth_fidelity/sync_numbers lambda-assignments -> def
- C408/C420/RUF005: dict() calls + concat -> literals / unpacking; dict-comp -> fromkeys
- RUF007: uncertainty ece bin loop -> itertools.pairwise
- RUF046: drop redundant int() around math.ceil/round
- PERF401: labels overlay_cmap append-loop -> list.extend
- SIM108/SIM300: ternary + un-yoda (readability-checked)

PLW0108: the 4 split test=lambda c: TESTSET.source(c) are NOT redundant wrappers
-- the lambda defers the module-global lookup so a testset monkeypatch/swap is
honored (the split-resolver tests replace the global); inlining froze early
binding and drifted the lock. Kept as lambda + bare noqa: PLW0108.

Full gate 0 errors; 754 tests green.

+++ core/data/dynamic/augment.py
@@ -82,7 +82,7 @@ class Augmentor:
     @staticmethod
     def _gaussian_kernel(sigma: float) -> torch.Tensor:
         """Separable 2D Gaussian kernel (sum=1), radius 3σ."""
-        r = max(1, int(math.ceil(3.0 * sigma)))
+        r = max(1, math.ceil(3.0 * sigma))
         xs = torch.arange(-r, r + 1, dtype=torch.float32)
         g = torch.exp(-(xs ** 2) / (2.0 * sigma * sigma))
         g = g / g.sum()
